+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- ...
+- Statement parameters of type OffsetDateTime and ZonedDateTime are converted into Timestamp.
 
 ### Changed
 - ...